requirements.txt
cachelib==0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Session==0.3.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
python-dotenv==0.13.0
requests==2.24.0
SQLAlchemy==1.3.17
urllib3==1.25.10
Werkzeug==1.0.1
